[] change in division logic.

+++ Calculator/Division.h
@@ -2,15 +2,15 @@
 #include "ArithmeticEquation.h"
 
 
-class Division : public ArithmeticEquation<int> {
+class Division : public ArithmeticEquation<float> {
 private:
-	int value1, value2;
+	float value1, value2;
 public:
-	Division(int value1, int value2);
-	void setValue1(int a);
-	void setValue2(int b);
-	int getValue1();
-	int getValue2();
-	int calculate();
+	Division(float value1, float value2);
+	void setValue1(float a);
+	void setValue2(float b);
+	float getValue1();
+	float getValue2();
+	float calculate();
 protected:
 };
